refactor: move user data & post data DTOs from `global-dto` folder
- Move user data dto from `global-dto` folder to `users/dto` folder.
- Move post data dto from `global-dto` folder to `posts/dto` folder.
