+++ src/posts/posts.resolver.ts
@@ -7,11 +7,11 @@ import {
   Parent,
 } from '@nestjs/graphql';
 import { PostsService } from './posts.service';
-import { Post } from '../global-dto/post-data.type';
+import { Post } from './dto/post-data.type';
 import { CreatePostInput } from './dto/create-post.input';
 import { EditPostInput } from './dto/update-post.input';
 import { SearchForPostInput } from './dto/search-for-post.input';
-import { User } from '../global-dto/user-data.type';
+import { User } from '../users/dto/user-data.type';
 import { UsersService } from '../users/users.service';
 import { PaginationOptions } from '../global/dto/pagination-options.dto';
 import { Types } from 'mongoose';
