+++ src/users/users.resolver.ts
@@ -10,8 +10,8 @@ import { UsersService } from './users.service';
 import { UpdateUserDataInput } from './dto/update-data-user.input';
 import { PostsService } from '../posts/posts.service';
 import { SearchForUserInput } from './dto/search-for-users.input';
-import { User } from '../global-dto/user-data.type';
-import { Post } from '../global-dto/post-data.type';
+import { User } from './dto/user-data.type';
+import { Post } from '../posts/dto/post-data.type';
 import { UserDocument } from './schemas/user.schema';
 import { PaginationOptions } from '../global/dto/pagination-options.dto';
 import { MongoObjectIdScalar } from '../global/dto/mongoObjectId.scalar';
